+++ lib/join/internal/diagnostic/diagnostic.go
